Introduce service for trimming config
