+++ src/sentry/relay/config/trimming/__init__.py
@@ -1,0 +1,10 @@
+from django.conf import settings
+
+from sentry.utils.services import LazyServiceWrapper
+
+from .base import Trimming
+
+backend = LazyServiceWrapper(
+    Trimming, settings.SENTRY_RELAY_TRIMMING, settings.SENTRY_RELAY_TRIMMING_OPTIONS
+)
+backend.expose(locals())
