+++ tests/sentry/api/endpoints/test_relay_projectconfigs.py
@@ -14,12 +14,9 @@ from sentry import quotas
 from sentry.constants import DataCategory, ObjectStatus
 from sentry.models.project import Project
 from sentry.models.relay import Relay
-from sentry.quotas.base import (
-    RETENTIONS_CONFIG_MAPPING,
-    RetentionSettings,
-    TrimmingConfig,
-    TrimmingConfigs,
-)
+from sentry.quotas.base import RETENTIONS_CONFIG_MAPPING, RetentionSettings
+from sentry.relay.config import trimming
+from sentry.relay.config.trimming.base import TrimmingConfig, TrimmingConfigs
 from sentry.testutils.helpers import Feature
 from sentry.testutils.pytest.fixtures import django_db_all
 from sentry.utils import safe
@@ -171,7 +168,9 @@ def test_internal_relays_should_receive_full_configs(
     else:
         assert safe.get_path(cfg, "config", "retentions") is None
 
-    trimming_configs = quotas.backend.get_trimming_configs(default_project.organization).to_object()
+    trimming_configs = trimming.backend.get_trimming_configs(
+        default_project.organization
+    ).to_object()
     if trimming_configs:
         assert safe.get_path(cfg, "config", "trimming") == trimming_configs
     else:
@@ -233,7 +232,7 @@ def test_parse_retentions_with_transactions(call_endpoint, default_project):
 @django_db_all
 def test_parse_trimming(call_endpoint, default_project):
     with patch.object(
-        quotas.backend,
+        trimming.backend,
         "get_trimming_configs",
         return_value=TrimmingConfigs(span=TrimmingConfig(max_size=17)),
     ):
