+++ tests/sentry/api/endpoints/test_relay_projectconfigs_v2.py
@@ -12,6 +12,7 @@ from sentry import quotas
 from sentry.constants import ObjectStatus
 from sentry.models.projectkey import ProjectKey, ProjectKeyStatus
 from sentry.quotas.base import RETENTIONS_CONFIG_MAPPING
+from sentry.relay.config import trimming
 from sentry.testutils.helpers import Feature
 from sentry.testutils.pytest.fixtures import django_db_all
 from sentry.utils import safe
@@ -122,7 +123,9 @@ def test_internal_relays_should_receive_full_configs(
     else:
         assert safe.get_path(cfg, "config", "retentions") is None
 
-    trimming_configs = quotas.backend.get_trimming_configs(default_project.organization).to_object()
+    trimming_configs = trimming.backend.get_trimming_configs(
+        default_project.organization
+    ).to_object()
     if trimming_configs:
         assert safe.get_path(cfg, "config", "trimming") == trimming_configs
     else:
